% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{BetaParams}
\alias{BetaParams}
\title{Generates beta prior from 95\% Confidence interval}
\usage{
BetaParams(mu, lower, upper)
}
\arguments{
\item{mu}{Mean of confidence interval}

\item{lower}{Upper bound of 95\% confidence interval}

\item{upper}{Upper bound of 95\% confidence interval}
}
\value{
Renders parameters of beta prior distribution
}
\description{
Generates beta prior from 95\% Confidence interval
}
\examples{
BetaParams(mu=0.5, lower=0.35, upper=0.72)
}
