% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Title Uses labels as variable names}
\usage{
clean_colnames(data)
}
\arguments{
\item{data}{Data to be cleaned}
}
\value{
Data with cleaned colnames
}
\description{
Title Uses labels as variable names
}
