% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day_length.R
\name{day_length}
\alias{day_length}
\title{Calculate Day Length}
\usage{
day_length(lat, date)
}
\arguments{
\item{lat}{Numeric vector. Latitude in degrees. Use positive values for north, negative for south.}

\item{date}{A Date object or a date-like string vector that can be coerced to Date.}
}
\value{
Numeric vector. Day length in hours.
}
\description{
\code{day_length} calculates the maximum possible day length (sunshine duration) based on latitude and date.
}
\examples{
day_length(lat = 35, date = as.Date("2023-03-15"))
day_length(lat = c(-15, 45), date = c("2024-06-21", "2023-12-21"))

}
