% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/declination.R
\name{declination}
\alias{declination}
\title{Solar Declination Calculation}
\usage{
declination(date, in_degrees = FALSE)
}
\arguments{
\item{in_degrees}{Logical. If \code{TRUE}, returns the declination in degrees. If \code{FALSE} (default), returns in radians.}

\item{dat}{A Date object, a date-like string, or a numeric value that can be coerced to Date.}
}
\value{
A numeric vector representing the solar declination angle in radians (default) or degrees.
}
\description{
\code{declination} calculates the solar declination (delta) for a given date.
}
\examples{
declination(as.Date("2024-05-03")) # Returns in radians (default)
declination("2024-05-03", TRUE) # Returns in degrees
declination(75) # Input Julian day
# Handling a large dataset of dates (e.g., 100,000 dates)
large_dates <- as.Date("2024-01-01") + 0:99999
result_large <- declination(large_dates)

}
\references{
Allen, R. G., Pereira, L. S., Raes, D., & Smith, M. (1998).
\emph{Crop Evapotranspiration: Guidelines for Computing Crop Water Requirements}.
FAO Irrigation and drainage paper 56. Food and Agriculture Organization of the United Nations, Rome.
}
\seealso{
\code{\link{doy}}
}
