% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doy.R
\name{doy}
\alias{doy}
\title{Calculate Day of the Year (DOY)}
\usage{
doy(dat)
}
\arguments{
\item{dat}{A date-like object or numeric input. This can be a \code{Date} object, a date-like string, or a numeric value representing Julian days.}
}
\value{
A numeric vector representing the day(s) of the year. If the input contains invalid date formats, an error will be thrown.
}
\description{
\code{doy} calculates the day of the year (DOY) for a given date. January 1st is represented as 1, and December 31st is 365 (or 366 in leap years).
}
\details{
If the input is numeric, the function treats it as Julian days. If the input is a \code{POSIXct} or \code{POSIXlt} object, it will be automatically coerced to a \code{Date} object.

The function uses \code{\link[base]{format}} to extract the day of the year from valid date objects.
}
\examples{
doy("2024-05-03") # Using a date-like string
doy(as.Date("2024-05-03 13:14:00")) # Using a Date object with time
doy(c("2023-01-01", "2023-12-31")) # Vectorized input with multiple dates
doy(150) # Using a numeric value (Julian day)

}
\seealso{
\code{\link[base]{as.Date}}, \code{\link[base]{format}}
}
