% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{eight_times_table}
\alias{eight_times_table}
\title{Title gives 8 times table}
\usage{
eight_times_table(x)
}
\arguments{
\item{x}{input to the function}
}
\value{
the product of 8 and another number
}
\description{
Title gives 8 times table
}
