% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{global_moranI}
\alias{global_moranI}
\title{Title}
\usage{
global_moranI(data, sf_object)
}
\arguments{
\item{data}{a numeric vector the same length as the neighbours list in sf object}

\item{sf_object}{spatial data as sf class}
}
\value{
Global moran value with p-value
}
\description{
Title
}
