% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{interpolate_population_agesex}
\alias{interpolate_population_agesex}
\title{Function to conduct  logistic population interpolation at time points with missing data.}
\usage{
interpolate_population_agesex(population_age_sex)
}
\arguments{
\item{population_age_sex}{data frame containing space, age, sex, year stratified data}
}
\value{
data frame with imputed (interpolated data)
}
\description{
Function to conduct  logistic population interpolation at time points with missing data.
}
\examples{
interpolate_population_agesex(naomi::demo_population_agesex)
}
