% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{log_linear_pop_agesex_interpolation}
\alias{log_linear_pop_agesex_interpolation}
\title{Function to conduct logistic population interpolation at time points with missing data.}
\usage{
log_linear_pop_agesex_interpolation(population_age_sex)
}
\arguments{
\item{population_age_sex}{data frame containing space, age, sex, year stratified data}
}
\value{
data frame with imputed (interpolated data)
}
\description{
Function to conduct logistic population interpolation at time points with missing data.
}
\examples{
log_linear_pop_agesex_interpolation(naomi::demo_population_agesex)
}
