% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{logNormal}
\alias{logNormal}
\title{Converts 95\% confidence interval to lognormal distribution}
\usage{
logNormal(mu, lowerCI, upperCI)
}
\arguments{
\item{mu}{Mean of confidence interval}

\item{lowerCI}{A lower bound of 95\% confidence interval}

\item{upperCI}{Upper bound of 95\% confidence interval}
}
\value{
Mean and sd of lognormal distribution
}
\description{
Converts 95\% confidence interval to lognormal distribution
}
\examples{
logNormal(0.02, 0.013, 0.035)
}
