% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{moran_clusters}
\alias{moran_clusters}
\title{Title}
\usage{
moran_clusters(data, sf_object)
}
\arguments{
\item{data}{vector that records data inputs from all areal units}

\item{sf_object}{shape file in sf class}
}
\value{
ggplot of clusters
}
\description{
Title
}
