% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{output_fixed_effects}
\alias{output_fixed_effects}
\title{Title}
\usage{
output_fixed_effects(inla_fit, tmb_fit)
}
\arguments{
\item{inla_fit}{Model outputs from INLA}

\item{tmb_fit}{Model outputs from INLA}
}
\description{
Title
}
