% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{rmvnorm_sparseprec}
\alias{rmvnorm_sparseprec}
\title{Title}
\usage{
rmvnorm_sparseprec(n, mean = rep(0, nrow(prec)), prec = diag(lenth(mean)))
}
\arguments{
\item{n}{Number of posterior samples}

\item{prec}{}
}
\description{
Title
}
