% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{run_inla}
\alias{run_inla}
\title{Title}
\usage{
run_inla(formula, data, verbose = TRUE)
}
\arguments{
\item{formula}{Formula for INLA}

\item{data}{data.frame with response variable and predictors}

\item{verbose}{}
}
\description{
Title
}
