% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{run_tmb}
\alias{run_tmb}
\title{Title}
\usage{
run_tmb(tmbdata, tmbpar, random_pars, dll)
}
\arguments{
\item{tmbdata}{List data inputs to TMB}

\item{tmbpar}{list of parameters to TMB}

\item{random_pars}{Random parameters}

\item{dll}{Dynamic library}
}
\description{
Title
}
