% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sol_rad.R
\name{sol_rad}
\alias{sol_rad}
\title{Estimate Solar Radiation using the Angstrom-Prescott model}
\usage{
sol_rad(
  lat,
  date,
  ssd,
  N = NULL,
  ext_rad = NULL,
  method = "standard",
  A = NA,
  B = NA
)
}
\arguments{
\item{lat}{Numeric. Latitude in degrees. Use positive values for north, negative for south.}

\item{date}{Date or character. A Date object or a date-like string that can be coerced to Date.}

\item{ssd}{Numeric. Sunshine duration in hours.}

\item{N}{Numeric. Numeric value for maximum possible sunshine duration (day length) in hours. If not provided, it is calculated.}

\item{ext_rad}{Optional. Numeric value for extraterrestrial radiation in MJ/m²/day. If not provided, it is calculated.}

\item{method}{Character. Calculation method: "standard" (Default), "parallel", or "gpu".}

\item{B}{Numeric. Angstrom-Prescott coefficient B (default is derived if not provided).}
}
\value{
Numeric. Estimated solar radiation in MJ/m²/day.
}
\description{
\code{sol_rad} estimates daily solar radiation on a horizontal surface based on sunshine duration using the Angstrom-Prescott model.
If extraterrestrial radiation (\code{ext_rad}) or maximum possible sunshine duration (\code{N}) is not provided, they are calculated within the function.
}
\examples{
sol_rad(lat = 35.0, date = as.Date("2023-03-15"), ssd = 8)
sol_rad(lat = -15.0, date = "2024-06-21", ssd = 10, method = "parallel")
sol_rad(lat = c(35,20), date = c("2023-03-15","2022-02-15"), ssd = c(8,5))

# Create a test vector with 1000 elements
set.seed(123)
test_lat <- runif(10000, -90, 90)
test_date <- as.Date("2023-01-01") + sample(1:365, 10000, replace = TRUE)
test_ssd <- runif(10000, 0, 24)
# Compare computation times
system.time({sol_rad(lat = test_lat, date = test_date, ssd = test_ssd, method = "standard")})
system.time({  sol_rad(lat = test_lat, date = test_date, ssd = test_ssd, method = "parallel")})

}
