% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{square}
\alias{square}
\title{Title - This function squares a number , any number}
\usage{
square(x)
}
\arguments{
\item{x}{inputs to the function}
}
\value{
out output from the model
}
\description{
Title - This function squares a number , any number
}
