% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{eight_times_table}
\alias{eight_times_table}
\title{8 times table}
\usage{
eight_times_table(x)
}
\arguments{
\item{x}{An input to the function}
}
\value{
The eight times  \code{x}.
}
\description{
8 times table
}
